Fix sca problems for aoe - phase 4 (#2272)
* fix sca problems for layout

* fix sca problems for metadata

+++ pkg/vm/engine/aoe/storage/buffer/pool.go
@@ -77,7 +77,7 @@ func (pool *SimpleMemoryPool) Free(size uint64) {
 	if size == 0 {
 		return
 	}
-	usagesize := atomic.AddUint64(&(pool.Usage), ^uint64(size-1))
+	usagesize := atomic.AddUint64(&(pool.Usage), ^(size - 1))
 	if usagesize > pool.Capacity {
 		panic("")
 	}
